+++ test/leipzig/test/melody.clj
@@ -7,9 +7,9 @@
   (->> [{:time 90}] (where :time (bpm 90))) =>
     [{:time 60}])
 
-(fact
-  (->> [{}] (where :part (is :bass))) =>
-    [{:part :bass}])
+(fact "where ignores nils."
+  (->> (rhythm [1 2]) (where :pitch dec)) =>
+    [{:time 0 :duration 1} {:time 1 :duration 2}])
 
 (fact "wherever can be used to provide default values to keys."
   (->> [{:time 0} {:time 1, :part :piano}]
@@ -34,6 +34,7 @@
   
   (phrase [1 1 2] [3 nil 4]) =>
     [{:time 0 :duration 1 :pitch 3}
+     {:time 1 :duration 1}
      {:time 2 :duration 2 :pitch 4}]
   
   (phrase [1 2] [0 [2 4]]) =>
@@ -61,10 +62,7 @@
 (fact
   (->> (phrase [1] [2]) (then (phrase [3] [4]))) =>
     [{:time 0 :duration 1 :pitch 2}
-     {:time 1 :duration 3 :pitch 4}]
-  (->> (phrase [1] [2]) (then 4 (phrase [3] [4]))) =>
-    [{:time 0 :duration 1 :pitch 2}
-     {:time 4 :duration 3 :pitch 4}])
+     {:time 1 :duration 3 :pitch 4}])
 
 (fact
   (->> (phrase [1] [2]) (then (after -2 (phrase [3 1] [4 5])))) =>
@@ -83,10 +81,7 @@
 (fact
   (->> (phrase [2] [1]) (times 2)) =>
     [{:time 0 :duration 2 :pitch 1}
-     {:time 2 :duration 2 :pitch 1}]
-  (->> (phrase [2] [1]) (times 2 4)) =>
-    [{:time 0 :duration 2 :pitch 1}
-     {:time 4 :duration 2 :pitch 1}])
+     {:time 2 :duration 2 :pitch 1}])
 
 (fact
   (->> (phrase [1 2] [2 3]) (after 1) (with (phrase [2] [1]))) =>