+++ CHANGELOG.md
@@ -6,6 +6,9 @@ document them here.
 
 New in 0.8.0
 ------------
+* *Breaking change* - `where` ignores nils.
+* *Breaking change* - In `phrase` rests are reified as notes of nil pitch.
+* *Breaking change* - `then` and `times` do not allow padding.
 * *Breaking change* - `bpm` returns a fn that converts beats to seconds, to match Supercollider. 
 * *Addition* - `mapthen` provides a temporal equivalent to mapcat.